+++ src/lib/services/films/getAll.ts
@@ -1,4 +1,4 @@
-import { Film } from '../../../types'
+import { Film } from '../../types'
 import { httpClient } from '../httpClient'
 
 type FilmResponse = Film[]
